Adding team stats
Added the full team as a player class so I can reuse the stats calculations methods.

Started to add dictionaries for the stats to be recorded while each player is on the bench, and then compared to non-bench scores. Decided to save at this point in case I mess up the next bit